Add a new empty line only if the file did not end with an empty line. refs #7
During the testing, I realized that the program adds a new empty line at the end of the file even if the file already ends with an empty line.
The root cause was that the EmptyLineAtEnd always add a new line at the end of the file, assuming that the format() will be called only if the file does not end with an empty line. However, the program will call format() from all rules if the file could not pass any check(). In other words, the format() from EmptyLineAtEnd will be called if the file violates other rules.
As a solution, I added a condition to EmptyLineAtEnd.format() that it only appends a new line if the file does not end with an empty line at the end.