Chain the formatting executions. refs #8
During the testing, I realized that the CommentBeginsWithSpace.format() is not applied.
The root cause was that the Rules.format() did not chain the Rule.format() correctly. Instead, it was applying the initial source code to all rules.
As a solution, I pass the result of previous Rule.format() to the next Rule.